+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## TBD 
+### Added
+- Filtering of the SNS messages produced in ApiarySNSListener using the `eventType` Message Attribute.
+
 ## [1.2.0] - 2019-03-13
 ### Added
 - Old locations for Table and Partitions to ApiarySNSListener events.